VIRO-1009: video callbacks not working
The weak reference change for ViroVideo resulted
in the video delegates getting GCed too soon.
Fix this by keeping a strong reference to the JNI
delegate in the video component.

Also remove the inner class; even though it may
work with the GC in this case, the memory relationships
are more clear with a static class.


Former-commit-id: 94b1c35c4894d88e5eb205d6279a86bb8b73b50e
Former-commit-id: 43ad0f858a6e85c3b35cc0184275f221403c575f