VIRO-927: Exit VR Callback Broke Android Exit Behavior
- Overriding the Android "X" Button callback resulted in a bug where
a developer with a normal template viro scene, with no onViroExit
implementation, would not be able to quit the activity (because
we overwrote the back behavior).

Thus, fix is to default to GVR's behavior of finishing the activity
if no such callbacks are applied in the SceneNavigator js file. Else,
we trigger the developer's desired callbacks.


Former-commit-id: 67b3346c2d7a1a0a73f2172bdc9c0628cb9887f3
Former-commit-id: 3c0b8a02e8f104225a2ff8fd4b9ac46b3234e85a