VIRO-678: Node Positional Update API (Bridge)
- Add onTransformUpdate callback prop to notify all JS node controls
whenever a positional update has occured. Updates only occur if the
callback prop is set.

- Developers can now set positions on the object without "resetting"
the positional props of the object to flush out inital props. This
is because we compare the control's native position to the incoming
prop, and then forcing and update if the native position differs.

- Updated both Android and iOS native bridge managers to include
the change.


Former-commit-id: a94ebb70d01b677be4189778044a1cc39b72a07c
Former-commit-id: 5b97123647da87b2a481430fbdb8a5ae81c18e6f