VIRO-1830: deleteGL method to clean up VROViewAR
The CVOpenGLTextureCache was deadlocking on its
destructor because it was being destroyed alongside
the VROViewAR, which shares the same EAGLContext
as the CVOpenGLTextureCache. The deadlock was
occurring internal to Apple so it’s unclear what was
actually causing the deadlock.

Fixed this by deleting the EAGLContext dependent
resources *prior* to the destruction of VROViewAR
via the new deleteGL method.


Former-commit-id: b1dbec7c6412c45650ed90cc4ac6e1f81a16f921
Former-commit-id: 2a130f70836d7d71418eac66ebccb67a5fb7d23b