VIRO-3320 ViroARImageMarker crashes the app
We weren't seeing the crash because our internet is fast. The issue
was due to ARImageTargets being loaded asynchronously such that when
the ARImageMarker was created, the targets were still being created.
When the targets finally load, they called the completion handler that
the ARImageMarker gave on a background thread and somewhere along the
creation of the native ARImageMarker, it would call setHidden which
has a passert_thread()


Former-commit-id: 32401de24655a51b2fab5c0e8081058c98b38c9b
Former-commit-id: cbfca3d22de5bf210ccac8a14b3e2ab0b49b1869