VIRO-4038: duration override in React API
Adds ability to override the duration of a skeletal
or keyframe animation. Because this only applies
to those types of animations, we only include this
property in Viro3DObject, not in other components.
Other components should use the animation’s set
duration; or we can consider making duration
override universal (e.g. also make it work for
registered animations).


Former-commit-id: 3f4ab05342ddab0ee7a32c3d76d20cedba6e37ad
Former-commit-id: 6c17ff2fc149477308b61063396f16aaa5af4f52