VIRO-4302 [iOS] Sounds don't get onFinish callbacks when loop=true
In the renderer, we were setting AVAudioPlayer's numberOfLoops to -1 in order to loop the
audio. Side effect was, AVAudioPlayer didnt trigger the sound finished
callback in that case. Fixed it by not setting it in that class, but instead
handling looping in bridge manually (by seeking to 0 and replaying) after
firing the onFinish call back


Former-commit-id: 2209f5c14a9477ae6bbeacc92be3a28e5e6b7a37
Former-commit-id: 6facb039f6cffb3d68fe189df22a79cc04ccbebd